Move scikit-learn interface to existing ekf interface
